refactor: move platform/index to platform.ts
