+++ packages/log/ts/console.ts
@@ -1,6 +1,6 @@
 import { required } from 'type-plus'
 import { logLevels } from './logLevels.js'
-import { isConsoleDebugAvailable } from './platform/index.js'
+import { isConsoleDebugAvailable } from './platform.js'
 import { plainLogFormatter } from './reporter.js'
 import type { LogFormatter, LogReporter, LogReporterOptions } from './types.js'
