+++ packages/log/ts/platform.browser.ts
@@ -1,4 +1,4 @@
-import type { LogEntry } from '../types.js'
+import type { LogEntry } from './types.js'
 
 export const toInspectLogEntry = (e: LogEntry) => e
 export function isConsoleDebugAvailable() {
