+++ packages/log/ts/platform.ts
@@ -1,5 +1,5 @@
 import { inspect } from 'util'
-import type { LogEntry } from '../types.js'
+import type { LogEntry } from './types.js'
 
 export function semverGt(versionString: string, version: [number, number, number]) {
   const actual = versionString.split('.').reverse().reduce((p, v, i) => {
